코드리팩토링: view 컴포넌트 분리 작업
