+++ view/src/pages/auth/SignUp.jsx
@@ -2,48 +2,7 @@ import { useState, useEffect, useRef } from 'react'
 import Logo from '@assets/user/empty_user.png'
 import Alert from '@components/commons/Alert.jsx'
 import { POST, PUT } from '@utils/Network.js'
-
-const Step1 = ({auth, setAuth, authEvent, offEvent, timerRef, user}) => {
-  const interval = 180
-  const [counter, setCounter] = useState(interval)
-  const [show, isShow] = useState(false)
-  useEffect(() => {
-    offEvent(counter)
-    if(counter > 0) {
-      timerRef.current = setInterval(() => setCounter(pre => pre - 0.5), 500)
-      return () => clearInterval(timerRef.current)
-    }
-  }, [counter]);
-  const changeEvent = (e) => setAuth(e.target.value)
-  const emailEvent = () => {
-    isShow(true)
-    POST('/oauth/user/email', user).then(
-      (res) => {
-        if(res.status) {
-          setCounter(interval)
-        }
-        isShow(false)
-      },
-      (err) => {
-        console.error(err)
-        isShow(false)
-      }
-    )
-  }
-  return (
-    <div className="input-group mt-3">
-        <div className="form-floating flex-grow-1">
-          <input type="text" className="form-control" id="code" name="code" placeholder="인증코드" value={auth} onChange={changeEvent} />
-          <label htmlFor="code">인증코드</label>
-        </div>
-        {counter > 0 ?
-          <button type="button" className="btn btn-outline-primary" style={{width: '75px'}} onClick={authEvent}>{Math.floor(counter)}초</button>
-          :
-          <button type="button" className="btn btn-outline-primary" style={{width: '75px'}} onClick={emailEvent} disabled={show}>재전송</button>
-        }
-    </div>
-  )
-}
+import SignUpStep1 from '@components/auth/SignUpStep1.jsx'
 
 const SignUp = () => {
   const [user, setUser] = useState({email: '', name: '', type: 1})
@@ -147,7 +106,7 @@ const SignUp = () => {
                 </form>
                 {step == 4 && <Alert type={type} />}
 
-                {step >= 1 && step <= 2 && <Step1 auth={auth} setAuth={setAuth} authEvent={authEvent} offEvent={offEvent} timerRef={timerRef1} user={user} />}
+                {step >= 1 && step <= 2 && <SignUpStep1 auth={auth} setAuth={setAuth} authEvent={authEvent} offEvent={offEvent} timerRef={timerRef1} user={user} />}
                 {step == 2 && <Alert type={type} />}
                 {step == 3 && <Alert type={type} />}
                 <div className="d-flex btn-group mt-3 mb-2">
