+++ view/src/pages/item/List.jsx
@@ -2,6 +2,7 @@ import { useState, useEffect, useRef } from 'react'
 import EMPTY from '@assets/item/empty_img.jpg'
 import '@styles/card.css'
 import { GET } from '@utils/Network.js'
+import Pagination from '@components/commons/Pagination.jsx';
 
 const Items = ({items}) => {
   const baseUrl = import.meta.env.VITE_APP_GATEWAY_URL + '/oauth/file/i/'
@@ -35,55 +36,6 @@ const Items = ({items}) => {
   )
 }
 
-const Pagination = ({pagination, clickEvent, page, total}) => {
-  const oneStepEvent = type => {
-    if(type === 'p') {
-      clickEvent(page == 0 ? page : page - 1)
-    } 
-    if(type === 'n') {
-      clickEvent(page == total-1 ? page : page + 1)
-    }
-    if(type === 'P') {
-      clickEvent(0)
-    }
-    if(type === 'N') {
-      clickEvent(total-1)
-    }
-  }
-  return (
-    <div>
-      <ul className="pagination justify-content-center">
-        <li className="page-item">
-          <button className="page-link" aria-label="First" onClick={() => oneStepEvent('P')}>
-            <span aria-hidden="true">&laquo;</span>
-          </button>
-        </li>
-        <li className="page-item">
-          <button className="page-link" aria-label="Previous" onClick={() => oneStepEvent('p')}>
-            <span aria-hidden="true">&lsaquo;</span>
-          </button>
-        </li>
-        {
-          pagination?.map((v, i) => {
-            return (
-              <li className="page-item" key={i} ><button className={v.active ? 'page-link active' : 'page-link'}  onClick={()=> clickEvent(i)}>{v.page}</button></li>
-            )
-          })
-        }
-        <li className="page-item">
-          <button className="page-link" aria-label="Next" onClick={() => oneStepEvent('n')}>
-            <span aria-hidden="true">&rsaquo;</span>
-          </button>
-        </li>
-        <li className="page-item">
-          <button className="page-link" aria-label="Last" onClick={() => oneStepEvent('N')}>
-            <span aria-hidden="true">&raquo;</span>
-          </button>
-        </li>
-      </ul>
-    </div>
-  )
-}
 
 const List = () => {
   const [items, setItems] = useState([])
@@ -91,7 +43,9 @@ const List = () => {
   const [page, setPage] = useState(0)
   const [total, setTotal] = useState(4)
   const [show, isShow] = useState(false)
-  const selectRef = useRef(null);
+  const [type, setType] = useState(0)
+  const [no, setNo] = useState('')
+  const [name, setName] = useState('')
   const inputRef = useRef(null);
   const getData = (params) => {
     GET('/mfr/item', params).then(res => {
@@ -118,47 +72,70 @@ const List = () => {
     })
     setPagination(arr)
   }
-  const selectEvent = () => {
-    inputRef.current.value = ''
+  const selectEvent = (e) => {
+    setType(Number(e.target.value))
+    setNo(0)
+    setName('')
+    if(Number(e.target.value) === 0) getData({page})
   }
   const submitEvent = (e) => {
     e.preventDefault();
-    const params = { page }
-    if(selectRef.current.value == 1) params.no = inputRef.current.value
-    if(selectRef.current.value == 2) params.name = inputRef.current.value
+    const params = {page}
+    if(type === 1) params.no = no
+    if(type === 2) params.name = name
     getData(params)
   }
   useEffect(() => {getData({ page })}, [page])
   return (
     <section className="container" style={{minHeight: '70vh'}}>
       <div>
-        <h2 className="mb-4">품목 관리</h2>
+        <h2 className="mb-4">품목 목록</h2>
       </div>
       <div className="mb-4">
         <form className="form" onSubmit={submitEvent} >
           <div className="row">
             <div className="col-md-4">
-              <select className="form-select type-select" ref={selectRef} onChange={selectEvent} >
+              <select className="form-select type-select" onChange={selectEvent} >
                 <option value="0">전체</option>
                 <option value="1">품목번호</option>
                 <option value="2">품목명</option>
               </select>
             </div>
+            {type !== 0 &&
             <div className="col-md-8">
               <div className="input-group">
-                <input type="input" className="form-control" ref={inputRef} />
+                {type === 1 &&
+                <input type="input" className="form-control" name="no" onChange={e => setNo(e.target.value)}/>
+                }
+                {type === 2 &&
+                <input type="input" className="form-control" name="name" onChange={e => setName(e.target.value)} />
+                }
                 <button type="submit" className="btn btn-outline-success">검색</button>
               </div>
             </div>
+            }
           </div>
         </form>
       </div>
-      {show && 
-      <>
-        <Items items={items} />
-        <Pagination pagination={pagination} clickEvent={clickEvent} page={page} total={total} />
-      </>
-      }
+      {show && (
+        items.length > 0
+          ? (
+            <>
+              <Items items={items} />
+              <Pagination
+                pagination={pagination}
+                clickEvent={clickEvent}
+                page={page}
+                total={total}
+              />
+            </>
+          )
+          : (
+            <div className="text-center py-5">
+              <p className="mb-0">검색 결과가 없습니다.</p>
+            </div>
+          )
+      )}
   </section>
   )
 }
