+++ view/src/pages/user/Index.jsx
@@ -3,56 +3,7 @@ import EmptyUser from '@assets/user/empty_user.png'
 import '@styles/user/user.css'
 import '@styles/card.css'
 import { GET, PATCH, DELETE } from '@utils/Network.js'
-
-const Pagination = ({pagination, clickEvent, page, total}) => {
-  const oneStepEvent = type => {
-    if(type === 'p') {
-      clickEvent(page == 0 ? page : page - 1)
-    } 
-    if(type === 'n') {
-      clickEvent(page == total-1 ? page : page + 1)
-    }
-    if(type === 'P') {
-      clickEvent(0)
-    }
-    if(type === 'N') {
-      clickEvent(total-1)
-    }
-  }
-  return (
-    <div>
-      <ul className="pagination justify-content-center">
-        <li className="page-item">
-          <button className="page-link" aria-label="First" onClick={() => oneStepEvent('P')}>
-            <span aria-hidden="true">&laquo;</span>
-          </button>
-        </li>
-        <li className="page-item">
-          <button className="page-link" aria-label="Previous" onClick={() => oneStepEvent('p')}>
-            <span aria-hidden="true">&lsaquo;</span>
-          </button>
-        </li>
-        {
-          pagination?.map((v, i) => {
-            return (
-              <li className="page-item" key={i} ><button className={v.active ? 'page-link active' : 'page-link'}  onClick={()=> clickEvent(i)}>{v.page}</button></li>
-            )
-          })
-        }
-        <li className="page-item">
-          <button className="page-link" aria-label="Next" onClick={() => oneStepEvent('n')}>
-            <span aria-hidden="true">&rsaquo;</span>
-          </button>
-        </li>
-        <li className="page-item">
-          <button className="page-link" aria-label="Last" onClick={() => oneStepEvent('N')}>
-            <span aria-hidden="true">&raquo;</span>
-          </button>
-        </li>
-      </ul>
-    </div>
-  )
-}
+import Pagination from '@components/commons/Pagination.jsx';
 
 const User = () => {
   const [depts, setDepts] = useState(null)
@@ -103,7 +54,6 @@ const User = () => {
     setDepts(data?.depts?.map(d => d.deptName).join(', '))
     changeLicence(data)
     isInfo(true)
-    console.log(data)
   }
   const submitEvent = e => {
     e.preventDefault()
